Add cpu tracing
This keeps track of the before and after state of each register/memory
access that an instruction touchs.  This is done by tapping into register/memory
reads and writes that the emulator is doing for a given instruction.  Writes
to memory and system registers will also trigger a re-read of the data
after its been written.

cpu tracing must be enabled at compile time with --features cpu-trace, which
will enabled the cputrace command in the debugger.  When cputrace is enabled
the step command will include the trace data.

(vb-rs 0xfff00160) > s
  0xfff00160  41d10000    st.b 0[r1], r10
             PSW: 0x0000400a => 0x0000400a
        disp i16: 0x0000 (0)
        read r01: 0x02000020
        read r10: 0x00000019
     write mem.b: addr: 0x02000020, before: 0x02, written: 0x19, after: 0x1b
        write PC: 0xfff00160 => 0xfff00164 (4)
      cpu cycles: 4
(vb-rs 0xfff00164) > s
  0xfff00164  0f4d        cmp 15, r8
             PSW: 0x0000400a => 0x00004002
        read r08: 0xffffffba
          imm i5: 0x0f (15)
        write PC: 0xfff00164 => 0xfff00166 (2)
      cpu cycles: 1